코드 정리 및 코드 주석 처리

+++ Assets/Scripts/Playing/ActiveObject/Effect/EffectAction/ExtinctionAction/SpritePlayCompleteExtinction.cs
@@ -1,33 +1,36 @@
 using System.Collections;
 using UnityEngine;
 
+
+// 스프라이트가 모두 재생되면 비활성화
 public class SpritePlayCompleteExtinction : EffectAction, IEffectExtinctionAction
 {
-    private int count;
+    // 재생될 스프라이트 개수
+    private byte count;
     protected override void OnInvoke()
     {
-        try
-        {
+        // ISpriteChangeAction에서 스프라이트 개수 가져오기
+        try{
             count = GetComponent<ISpriteChangeAction>().spriteCount;
-
         }
         catch (System.Exception ex)
         {
             Debug.LogError($"에러 발생! {ex}");
         }
 
+        // 스트라이트 재생을 간접적으로 기록하기 위해 코루틴 시작
         StartCoroutine(startExtinctionTimer());
-
     }
 
+    // 스프라이트 재생이 모두 끝났는지 예측하는 코루틴
     IEnumerator startExtinctionTimer()
     {
-
+        // 스프라이트 재생이 모두 끝날 때까지 대기
         while (count > 0)
         {
             count--;
-
             float time = 0f;
+
             while (time <= 1f)
             {
                 time += Time.deltaTime * Utils.EFFECT_SPEED;
@@ -35,6 +38,7 @@ public class SpritePlayCompleteExtinction : EffectAction, IEffectExtinctionActio
             }
         }
 
+        // 재생이 모두 끝났으면 이펙트 비활성화
         effect.Disable();
     }
 }
